Fix datatype mismatch for quant_dot_kernel.

+++ test_puzzle.py
@@ -28,7 +28,7 @@ def test(puzzle, puzzle_spec, nelem={}, B={"B0": 32}, print_log=False, device="c
         tt_args.append(torch.rand(*v, device=device) - 0.5)
         # tt_args.append(torch.ones(*v, device=device))
         if t is not None and t.annotation.dtype == "int32":
-            tt_args[-1] = torch.randint(-100000, 100000, v, device=device)
+            tt_args[-1] = torch.randint(-100000, 100000, v, dtype=torch.int32, device=device)
 
     grid = lambda meta: (triton.cdiv(nelem["N0"], meta["B0"]),
                          triton.cdiv(nelem.get("N1", 1), meta.get("B1", 1)),
